Split up commands and commons
