+++ src/commands/index.ts
@@ -1,0 +1,2 @@
+export { translateAllAndReplace } from '../commands/TranslateAllAndReplace';
+export { translateStringsAndReplace } from '../commands/TranslateStringsAndReplace';
